requirements.txt
# XAUUSD AI Trading System — Python Dependencies
# Install with: pip install -r requirements.txt

# Web API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# ML / Data
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
joblib==1.4.2

# Telegram bot
python-telegram-bot==21.2
watchdog==4.0.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.5
